requirements.txt
Flask
sentence-transformers
faiss-cpu
transformers
PyMuPDF
torch
